refactor(web): F4 主抄本 ingest 下沉 useTranscriptIngest hook
messages/draft + 7 个配对去重 ref + setMsgs/setDraftBoth/pushMsg/pushSystem/
commitDraft/pairToolResult/handleCli/applyHistory 逐字迁入 hooks/useTranscriptIngest.ts；
phase/initInfo/permMode 是 cli 流驱动的会话元数据，状态留 Chat、稳定 setter 注入 hook；
E3 历史加载 effect 留组合层（reset 先于建连），清空段前 7 步内聚为 api.reset()。
WS 事件分发（F5 前留 Chat）经 api 的 ref 出口读写。Chat.tsx 851 行。零行为改动。
